+++ net-deleter/src/handlers/clear_buffer_handler.rs
@@ -1,28 +1,26 @@
 use std::error::Error;
 use std::sync::Arc;
-
-use net_deleter_api::api::network::DeleteNetworkRequestDTO;
+use net_deleter_api::api::buffer::ClearBufferRequestDTO;
 use net_component::handler::network_service_handler::NetworkServiceHandler;
 use net_core_api::api::envelope::envelope::Envelope;
-use net_core_api::core::decoder_api::Decoder;
 use net_core_api::core::typed_api::Typed;
 use sqlx::Pool;
 use sqlx::Postgres;
 
 use crate::core::delete_error::DeleteError;
-use crate::utils::network_deleter;
+use crate::utils::buffer_cleaner;
 
 #[derive(Default, Debug)]
-pub struct DeleteNetworkHandler {}
+pub struct ClearBufferHandler {}
 
-impl DeleteNetworkHandler {
+impl ClearBufferHandler {
     pub fn boxed(self) -> Box<Self> {
         Box::new(self)
     }
 }
 
 #[async_trait::async_trait]
-impl NetworkServiceHandler for DeleteNetworkHandler {
+impl NetworkServiceHandler for ClearBufferHandler {
     async fn handle(&self, connection_pool: Arc<Pool<Postgres>>, enveloped_request: Envelope) -> Result<Envelope, Box<dyn Error + Send + Sync>> {
         let deletable_data_type = self.get_handler_type().split('-').collect::<Vec<_>>().join(" ");
         if enveloped_request.get_envelope_type() != self.get_handler_type() {
@@ -31,14 +29,12 @@ impl NetworkServiceHandler for DeleteNetworkHandler {
             ).into());
         }
         let tenant_id = enveloped_request.get_tenant_id();
-        let network_to_delete = DeleteNetworkRequestDTO::decode(enveloped_request.get_data());
         let mut transaction = match connection_pool.begin().await {
             Ok(transaction) => transaction,
             Err(err) => return Err(DeleteError::TranscationError(err.to_string()).into()),
         };
-        let delete_packets_res = network_deleter::delete_network_transaction(
+        let delete_packets_res = buffer_cleaner::clear_buffer_transaction(
             &mut transaction,
-            network_to_delete.get_id(),
             tenant_id,
         ).await;
         if let Err(err) = delete_packets_res {
@@ -49,6 +45,6 @@ impl NetworkServiceHandler for DeleteNetworkHandler {
     }
 
     fn get_handler_type(&self) -> String {
-        DeleteNetworkRequestDTO::get_data_type().to_string()
+        ClearBufferRequestDTO::get_data_type().to_string()
     }
 }