+++ net-reporter/src/handlers/chart_handlers/http_responses/handler.rs
@@ -59,6 +59,7 @@ const HTTP_RESPONSES_REQUEST_QUERY: &str = "
         AND Frametime >= $2
         AND Frametime < $3
         AND Http->'http.response.code' IS NOT NULL
+        AND Network_ID = $4
         {}
         {}
         {}
@@ -79,6 +80,7 @@ impl HttpResponsesHandler {
         tenant_id: &str,
         start_date: DateTime<Utc>,
         end_date: DateTime<Utc>,
+        network_id: &str,
         filters: &HttpResponsesFiltersDTO,
     ) -> Result<Vec<HttpResponseResponse>, Error> {
         log::info!("Query Parameters: {:?}", tenant_id);
@@ -88,6 +90,7 @@ impl HttpResponsesHandler {
             .add_param(tenant_id)
             .add_param(start_date)
             .add_param(end_date)
+            .add_param(network_id)
             .add_option_param(filters.is_include_http_methods_mode().map(|_| filters.get_http_responses().to_vec()))
             .add_option_param(filters.is_include_endpoints_mode().map(|_| filters.get_endpoints().to_vec()))
             .add_option_param(filters.get_bytes_lower_bound())
@@ -111,9 +114,10 @@ impl NetworkServiceHandler for HttpResponsesHandler {
         let request = HttpResponsesRequestDTO::decode(enveloped_request.get_data());
         let request_start_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_start_date_time()).unwrap();
         let request_end_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_end_date_time()).unwrap();
+        let network_id = request.get_network_id();
         let filters = request.get_filters();
 
-        let query = QueryBuilder::new(HTTP_RESPONSES_REQUEST_QUERY, 4)
+        let query = QueryBuilder::new(HTTP_RESPONSES_REQUEST_QUERY, 5)
             .add_dynamic_filter(filters.is_include_http_methods_mode(), 1, INCLUDE_HTTP_METHODS_FILTER_QUERY, EXCLUDE_HTTP_METHODS_FILTER_QUERY)
             .add_dynamic_filter(filters.is_include_endpoints_mode(), 1, INCLUDE_ENDPOINT_FILTER_QUERY, EXCLUDE_ENDPOINT_FILTER_QUERY)
             .add_static_filter(filters.get_bytes_lower_bound(), SET_LOWER_BYTES_BOUND, 1)
@@ -126,6 +130,7 @@ impl NetworkServiceHandler for HttpResponsesHandler {
             tenant_id,
             request_start_date,
             request_end_date,
+            network_id,
             filters,
         ).await?;