+++ net-reporter/src/handlers/chart_handlers/http_responses_distribution/handler.rs
@@ -46,6 +46,7 @@ const HTTP_RESPONSES_DIST_REQUEST_QUERY: &str = "
         Tenant_ID = $1
         AND Frametime >= $2
         AND Frametime < $3
+        AND Network_ID = $4
         AND Http->'http.response.code' IS NOT NULL
         {}
     GROUP BY
@@ -71,12 +72,14 @@ impl HttpResponsesDistributionHandler {
         tenant_id: &str,
         start_date: DateTime<Utc>,
         end_date: DateTime<Utc>,
+        network_id: &str,
         filters: &HttpResponsesDistributionFiltersDTO,
     ) -> Result<Vec<HttpResponsesDistributionBucketResponse>, Error> {
         SqlxQueryBuilderWrapper::<HttpResponsesDistributionBucketResponse>::new(query_string)
             .add_param(tenant_id)
             .add_param(start_date)
             .add_param(end_date)
+            .add_param(network_id)
             .add_option_param(filters.is_include_endpoints_mode().map(|_| filters.get_endpoints().to_vec()))
             .add_option_param(filters.get_bytes_lower_bound())
             .add_option_param(filters.get_bytes_upper_bound())
@@ -99,9 +102,10 @@ impl NetworkServiceHandler for HttpResponsesDistributionHandler {
         let request = HttpResponsesDistributionRequestDTO::decode(enveloped_request.get_data());
         let request_start_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_start_date_time()).unwrap();
         let request_end_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_end_date_time()).unwrap();
+        let network_id = request.get_network_id();
         let request_filters = request.get_filters();
 
-        let query = QueryBuilder::new(HTTP_RESPONSES_DIST_REQUEST_QUERY, 4)
+        let query = QueryBuilder::new(HTTP_RESPONSES_DIST_REQUEST_QUERY, 5)
             .add_dynamic_filter(request_filters.is_include_endpoints_mode(), 1, INCLUDE_ENDPOINT_FILTER_QUERY, EXCLUDE_ENDPOINT_FILTER_QUERY)
             .add_static_filter(request_filters.get_bytes_lower_bound(), SET_LOWER_BYTES_BOUND, 1)
             .add_static_filter(request_filters.get_bytes_upper_bound(), SET_UPPER_BYTES_BOUND, 1)
@@ -113,6 +117,7 @@ impl NetworkServiceHandler for HttpResponsesDistributionHandler {
             tenant_id,
             request_start_date,
             request_end_date,
+            network_id,
             request_filters,
         ).await?;