+++ net-reporter/src/handlers/chart_handlers/network_bandwidth_per_protocol/handler.rs
@@ -48,6 +48,7 @@ const NETWORK_BANDWIDTH_PER_PROTOCOL_REQUEST_QUERY: &str = "
         Tenant_ID = $1
         AND Frametime >= $2
         AND Frametime < $3
+        AND Network_ID = $4
         {}
     GROUP BY Separated_Protocols
     HAVING
@@ -70,12 +71,14 @@ impl NetworkBandwidthPerProtocolHandler {
         tenant_id: &str,
         start_date: DateTime<Utc>,
         end_date: DateTime<Utc>,
+        network_id: &str,
         filters: &NetworkBandwidthPerProtocolFiltersDTO,
     ) -> Result<Vec<ProtocolResponse>, Error> {
         SqlxQueryBuilderWrapper::<ProtocolResponse>::new(query_string)
             .add_param(tenant_id)
             .add_param(start_date)
             .add_param(end_date)
+            .add_param(network_id)
             .add_option_param(filters.is_include_endpoints_mode().map(|_| filters.get_endpoints().to_vec()))
             .add_option_param(filters.get_bytes_lower_bound())
             .add_option_param(filters.get_bytes_upper_bound())
@@ -98,9 +101,10 @@ impl NetworkServiceHandler for NetworkBandwidthPerProtocolHandler {
         let request = NetworkBandwidthPerProtocolRequestDTO::decode(enveloped_request.get_data());
         let request_start_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_start_date_time()).unwrap();
         let request_end_date: DateTime<Utc> = Utc.timestamp_millis_opt(request.get_end_date_time()).unwrap();
+        let network_id = request.get_network_id();
         let request_filters = request.get_filters();
 
-        let query = QueryBuilder::new(NETWORK_BANDWIDTH_PER_PROTOCOL_REQUEST_QUERY, 4)
+        let query = QueryBuilder::new(NETWORK_BANDWIDTH_PER_PROTOCOL_REQUEST_QUERY, 5)
             .add_dynamic_filter(request_filters.is_include_endpoints_mode(), 1, INCLUDE_ENDPOINT_FILTER_QUERY, EXCLUDE_ENDPOINT_FILTER_QUERY)
             .add_static_filter(request_filters.get_bytes_lower_bound(), SET_LOWER_BYTES_BOUND, 1)
             .add_static_filter(request_filters.get_bytes_upper_bound(), SET_UPPER_BYTES_BOUND, 1)
@@ -112,6 +116,7 @@ impl NetworkServiceHandler for NetworkBandwidthPerProtocolHandler {
             tenant_id,
             request_start_date,
             request_end_date,
+            network_id,
             request_filters,
         ).await?;